+++ src/browser-lib/google-classroom/googleClassroomApi.ts
@@ -173,6 +173,7 @@ export type AddOnContextResponse = {
     submissionId: string;
   };
   pupilLoginHint: string;
+  teacherLoginHint?: string | null;
 };
 export type AddOnContextArgs = {
   courseId: string;